+++ src/applications/disability-benefits/526EZ/form-entry.jsx
@@ -4,7 +4,7 @@ import './sass/disability-benefits.scss';
 import startApp from '../../../platform/startup';
 
 import routes from './routes';
-import reducer from './reducer';
+import reducer from './reducers';
 import manifest from './manifest.json';
 
 startApp({
