+++ src/applications/disability-benefits/526EZ/reducers/index.js
@@ -1,0 +1,9 @@
+import itf from './itf';
+import formConfig from '../config/form';
+
+import { createSaveInProgressFormReducer } from '../../../../platform/forms/save-in-progress/reducers';
+
+export default {
+  form: createSaveInProgressFormReducer(formConfig),
+  itf
+};
